+++ logAnalyzer.py
@@ -1,3 +1,4 @@
+# Copyright (C) 2022 Lucas Aimaretto / laimaretto@gmail.com
 # Copyright (C) 2020 Manuel Saldivar / manuelsaldivar@outlook.com.ar, Lucas Aimaretto / laimaretto@gmail.com
 #
 # This is logAnalyzer
@@ -21,6 +22,19 @@ import json
 import re
 from ttp import ttp
 
+DATA_VALUE     = 'Value'
+DATA_COMMAND   = '#Command:'
+DATA_MAJOR_DWN = '#majorDown:'
+DATA_FLTR_COLS = '#filterColumns:'
+DATA_FLTR_ACTN = '#filterAction:'
+
+RTR_ID = dict(
+	name = ['NAME'],
+	both = ['NAME','IP'],
+	ip   = ['IP']
+)
+
+
 def readTemplate(fileTemplate, templateFolder, templateEngine):
 	
 	# Read the list of templates passed by CSV of textFSM and return template read list (read)
@@ -38,9 +52,11 @@ def readTemplate(fileTemplate, templateFolder, templateEngine):
 	for i,tmpltName in enumerate(templates):
 
 		d[tmpltName] = {
-			'listOfcolumns':[],
+			'templateColumns':[],
 			'commandKey':'',
-			'majorDown':['down']
+			'majorDown':['down'],
+			'filterColumns':[],
+			'filterAction':None,
 		}	
 
 		fName = templateFolder+tmpltName
@@ -56,50 +72,127 @@ def readTemplate(fileTemplate, templateFolder, templateEngine):
 
 			if templateEngine == 'textFSM':
 
-				h1 = line.find('Value')
-				h2 = line.find('#Command:')
-				h3 = line.find('#majorDown:')
+				h1 = line.find(DATA_VALUE)
+				h2 = line.find(DATA_COMMAND)
+				h3 = line.find(DATA_MAJOR_DWN)
+				h4 = line.find(DATA_FLTR_COLS)
+				h5 = line.find(DATA_FLTR_ACTN)
 				
 				if h1 != -1:
+					# We identify here the variables
 					col = line.split(' ')[-2]
-					#listOfcolumns[i].append(col)
-					d[tmpltName]['listOfcolumns'].append(col)
+					d[tmpltName]['templateColumns'].append(col)
 				
 				if h2 != -1:
-					cmd = line.split(': ')[1].strip('\n')
-					#commandKey[i].append(cmd)
+					# Here we identify the command
+					line = line.replace(DATA_COMMAND + ' ', DATA_COMMAND)
+					cmd  = line.split(':')[1].strip('\n')
 					d[tmpltName]['commandKey'] = cmd
 
 				if h3 != -1:
-					keys = line.split(': ')[1].strip('\n').split(',')
+					# we add more major words to the list
+					line = line.replace(DATA_MAJOR_DWN + ' ', DATA_MAJOR_DWN)
+					keys = line.split(':')[1].strip('\n').split(',')
 					for key in keys:
 						d[tmpltName]['majorDown'].append(key)
+
+				if h4 != -1:
+					# We identify the columnes to be filtered
+					line = line.replace(DATA_FLTR_COLS + ' ', DATA_FLTR_COLS)
+					keys = line.split(':')[1].strip('\n').split(',')
+					for key in keys:
+						if key not in [None, '', ' ']:
+							d[tmpltName]['filterColumns'].append(key)
+
+				if h5 != -1:
+					# we identify the action to be performed on the filterd columns
+					line = line.replace(DATA_FLTR_ACTN + ' ', DATA_FLTR_ACTN)
+					action = line.split(':')[1].strip('\n')
+					d[tmpltName]['filterAction'] = action
 
 			if templateEngine == 'ttp':
 
-				h1 = line.find('Columns: ')
-				h2 = line.find('Command: ')
-				h3 = line.find('majorDown: ')
+				h1 = line.find('#Columns: ')
+				h2 = line.find('#Command: ')
+				h3 = line.find('#majorDown: ')
+				h4 = line.find('#filterColumns: ')
+				h5 = line.find('#filterAction: ')	
 				
 				if h1 != -1:
 					col = line.split(': ')[1].strip('\n').split(",")
-					#listOfcolumns[i].append(col)
-					d[tmpltName]['listOfcolumns'] = col
+					d[tmpltName]['templateColumns'] = col
 				
 				if h2 != -1:
 					cmd = line.split(': ')[1].strip('\n')
-					#commandKey[i].append(cmd)
 					d[tmpltName]['commandKey'] = cmd
 
 				if h3 != -1:
 					keys = line.split(': ')[1].strip('\n').split(',')
 					for key in keys:
 						d[tmpltName]['majorDown'].append(key)
+
+				if h4 != -1:
+					# We identify the columnes to be filtered
+					keys = line.split(': ')[1].strip('\n').split(',')
+					for key in keys:
+						if key not in [None, '', ' ']:
+							d[tmpltName]['filterColumns'].append(key)
+
+				if h5 != -1:
+					# we identify the action to be performed on the filterd columns
+					action = line.split(': ')[1].strip('\n')
+					d[tmpltName]['filterAction'] = action
+
+		if len(d[tmpltName]['filterColumns']) > 0:
+
+			print(f'The template {tmpltName} has the following columns to be filtered:')
+			print('Columns: '+str(d[tmpltName]['filterColumns']))
+
+			# checking column's names
+			x = [col for col in d[tmpltName]['filterColumns'] if col not in d[tmpltName]['templateColumns']]
+			if len(x) > 0:
+				print(f'There are some columns which are not original variables of the template.')
+				print(x)
+				print(f'Check the variables names. Quitting...')
+				quit()
+
+			# we want to filter columns
+			if d[tmpltName]['filterAction'] not in ['include-only','exclude']:
+				# we want to filter columns but we have not specified
+				# an action to perform
+				print(f'The the action to be used has not been properly set.')
+				print(f'Please set either "include" or "exclude" in the comments section of the template file.\nQuitting...')
+				quit()
+
+			if d[tmpltName]['filterAction'] == 'exclude':
+
+				# we check if the filter columns are equal to the templates' columns.
+				# if so, chance is we're getting an empty DF. We don't want this.
+				if sorted(d[tmpltName]['filterColumns']) == sorted(d[tmpltName]['templateColumns']):
+					print(f'The template {tmpltName} has the following columns:')
+					print(d[tmpltName]['templateColumns'])
+					print(f'Since the action to be performed is "exclude" all the columns will be filtered out and you')
+					print(f'will end with an empty table. Quitting...')
+					quit()
+
+				# since we are using 'exclude' our final list of filtered columns is the difference
+				# between the original 'templateColumns' and the 'filterColumns'
+				x = [col for col in d[tmpltName]['templateColumns'] if col not in d[tmpltName]['filterColumns']]
+				d[tmpltName]['filterColumns'] = x
+
+			#print('filterColumns: '+str(d[tmpltName]['filterColumns']))
+
+		else:
+			# if no filtering columns are defined, we assign those by the original
+			# template columns
+			d[tmpltName]['filterColumns'] = d[tmpltName]['templateColumns'].copy()
+			#print('filterColumns: '+str(d[tmpltName]['filterColumns']))
+
 
 	print(f'##### Successfully Loaded Templates from folder {templateFolder} #####')
 	return d 
 
-def makeParsed(nomTemplate, routerLog, templateFolder, templateEngine, columnss):
+def makeParsed(nomTemplate, routerLog, templateFolder, templateEngine, templateColumns):
 	"""
 	Parse through textFSM (reading the file again)
 
@@ -119,7 +212,7 @@ def makeParsed(nomTemplate, routerLog, templateFolder, templateEngine, columnss)
 		parsed_results   = results_template.ParseText (routerLog)
 
 		# With list of results, we build a Pandas DataFrame
-		parsed_results = pd.DataFrame(parsed_results, columns= columnss)
+		parsed_results = pd.DataFrame(parsed_results, columns= templateColumns)
 
 	if templateEngine == 'ttp':
 
@@ -132,7 +225,7 @@ def makeParsed(nomTemplate, routerLog, templateFolder, templateEngine, columnss)
 		output = parser.result(format='table')
 		parsed_results = output[0][1][0]
 
-		parsed_results = pd.DataFrame(parsed_results, columns= columnss)
+		parsed_results = pd.DataFrame(parsed_results, columns= templateColumns)
 
 	return parsed_results
 
@@ -184,14 +277,16 @@ def readLog(logFolder, formatJson):
 
 	return d
 
-def parseResults(dTmpl, dLog, templateFolder, templateEngine):
+def parseResults(dTmpl, dLog, templateFolder, templateEngine, routerId):
 	"""
 	Build the Dataframe from textFSM filter, index and router log
 
 	Args:
-		dTmpl (dict):        dictionary with info from templates.
-		dLog (dict):         dicitonary with logs. Each key is the fileName; the value, is the content of the log.
+		dTmpl (dict):           dictionary with info from templates.
+		dLog (dict):            dicitonary with logs. Each key is the fileName; the value, is the content of the log.
 		templateFolder (str):   folder of templates
+		templateEngine:         textFsm or ttp
+		routerId:               name, IP or both
 
 	Returns:
 		datosEquipo (dict): Dictionary where keys are templateNames. For each key, a DF with parsed results.
@@ -199,82 +294,79 @@ def parseResults(dTmpl, dLog, templateFolder, templateEngine):
 
 	datosEquipo  = {}
 
-	for tmpltName in dTmpl.keys():
+	for idT, tmpltName in enumerate(dTmpl.keys()):
 
-		columnss    = dTmpl[tmpltName]['listOfcolumns']
-		commandKey  = dTmpl[tmpltName]['commandKey']
-		dfTemp      = pd.DataFrame(columns=columnss)
+		templateColumns = dTmpl[tmpltName]['templateColumns']
+		commandKey   	= dTmpl[tmpltName]['commandKey']
+		filterCols   	= dTmpl[tmpltName]['filterColumns']
 
-		for routerLogKey in dLog.keys():
+		#if len(filterCols)>0:
+		orderedColums = RTR_ID[routerId] + filterCols
+		dfTemp        = pd.DataFrame(columns=orderedColums)
+		#else:
+		#	dfTemp       = pd.DataFrame(columns=templateColumns)
+
+		for idR, routerLogKey in enumerate(dLog.keys()):
 
 			routerLogFname  = routerLogKey.split("/")[-1]
 
-			print(routerLogFname , tmpltName)
+			print(idT, idR, tmpltName, routerLogFname)
 
-			if '_rx.json' in routerLogFname: 
-				# If text format is json, else, we continue work with rx_txt
-				# TODO: better way to detect json format.
+			routerName = dLog[routerLogKey]['name']
+			routerIP   = dLog[routerLogKey]['ip']
 
-				routerName = dLog[routerLogKey]['name']
+			# logs es cada comando que se ejecuto en el router, dentro del json file.
+			for cmdsLogs in dLog[routerLogKey].keys():
 
-				# logs es cada comando que se ejecuto en el router, dentro del json file.
-				for cmdsLogs in dLog[routerLogKey].keys():
+				# prog es el nombre del comando en cada template file
+				prog = re.compile(commandKey)
 
-					# prog es el nombre del comando en cada template file
-					prog = re.compile(commandKey)
+				# searchKey es el regex match entre logs y prog
+				match = prog.search(cmdsLogs)
 
-					# searchKey es el regex match entre logs y prog
-					match = prog.search(cmdsLogs)
+				if match: 
+					#if command(in template) == command(in key of router) then we stores log info in routeLog variable
+					# significa que el comando se ejecutó en el router y existe un template
+					# para ese comando.
 
-					if match: 
-						#if command(in template) == command(in key of router) then we stores log info in routeLog variable
-						# significa que el comando se ejecutó en el router y existe un template
-						# para ese comando.
+					# {
+					# 	'logs1':'output1',
+					# 	'logs2':'output2',
+					# 	'logsN':'outputN',
+					# }
 
-						# {
-						# 	'logs1':'output1',
-						# 	'logs2':'output2',
-						# 	'logsN':'outputN',
-						# }
+					# "/show router 4001 route-table | match No": "No. of Routes: 566",
+					# "/show router 4002 route-table | match No": "MINOR: CLI Invalid router \"4002\".\u0007",
+					# "/show router route-table | match No": "No. of Routes: 3337",						
 
-						# "/show router 4001 route-table | match No": "No. of Routes: 566",
-						# "/show router 4002 route-table | match No": "MINOR: CLI Invalid router \"4002\".\u0007",
-						# "/show router route-table | match No": "No. of Routes: 3337",						
+					routerLog = cmdsLogs + '\n' + dLog[routerLogKey][cmdsLogs] + '\n'
 
-						routerLog = cmdsLogs + '\n' + dLog[routerLogKey][cmdsLogs] + '\n'
+					# We parse results from the key:value association
+					# A list is returnd with results
+					# to parse, with provide the complete set of columns as defined inside the template: templateColumns
+					dfResult = makeParsed(tmpltName, routerLog, templateFolder, templateEngine, templateColumns)
 
-						# We parse results from the key:value association
-						# A list is returnd with results
-						dfResult = makeParsed(tmpltName, routerLog, templateFolder, templateEngine, columnss)
+					# If there are columns to be filtered, we reduced the 
+					# size of the DF to that number of columns
+					if len(filterCols) > 0:
+						dfResult = dfResult[filterCols]
+
+					# We need to define the identification of the router.
+					if routerId == 'name':
+						dfResult['NAME'] = routerName
+
+					elif routerId == 'both':
 
 						dfResult['NAME'] = routerName
-						dfTemp = pd.concat([dfTemp, dfResult])
+						dfResult['IP']   = str(routerIP)
+					
+					elif routerId == 'ip':
 
-			else:
-				# if here, we analyze plain text Files
-				pass
-				# routerName = routers[i1][0].replace('_rx.txt','')
-				# routerLog  = content[i1]
+						dfResult['IP'] = str(routerIP)
 
-				# # "/show router 4001 route-table | match No": "No. of Routes: 566",
-				# # "/show router 4002 route-table | match No": "MINOR: CLI Invalid router \"4002\".\u0007",
-				# # "/show router route-table | match No": "No. of Routes: 3337",	
+					dfResult = dfResult[orderedColums]
 
-				# parsed_results = makeParsed(nomTemplate, routerLog)
-
-				# if len(parsed_results) == 0:
-				# 	# if the parse is empty, we save the name of the routers
-				# 	parsed_results = [routerName]
-				# 	for empty in range(len(columnss)-1):
-				# 		parsed_results.append('NOT VALUE')
-
-				# 	parsed_results = [parsed_results]
-				# 	dfResult = pd.DataFrame(parsed_results, columns= columnss)
-				# else:
-				# 	dfResult = pd.DataFrame(parsed_results, columns= columnss)
-				# 	dfResult['NAME'] = routerName
-
-				# dfTemp = pd.concat([dfTemp, dfResult])
+					dfTemp = pd.concat([dfTemp, dfResult])
 
 		# It is stored in the dataEquipment dictionary with the key nomTemplate
 		# the DF with the data of all routers
@@ -283,16 +375,18 @@ def parseResults(dTmpl, dLog, templateFolder, templateEngine):
 		# I added this here because it was already done in main ().
 		# It is cleaner like this ...
 		datosEquipo[tmpltName].reset_index(level=0, inplace=True)
-		datosEquipo[tmpltName] = datosEquipo[tmpltName].drop(columns='index')		
+		datosEquipo[tmpltName] = datosEquipo[tmpltName].drop(columns='index')
 
 	return datosEquipo
 
-def searchDiff(datosEquipoPre, datosEquipoPost, dTmplt):
+def searchDiff(datosEquipoPre, datosEquipoPost, dTmplt, routerId):
 	#Makes a new table, in which it brings the differences between two tables (post-pre)	
 	
-	countDif = {}	
+	countDif   = {}
 
 	for tmpltName in datosEquipoPre.keys():
+
+		filterCols = dTmplt[tmpltName]['filterColumns']
 
 		dfUnion = pd.merge(datosEquipoPre[tmpltName], datosEquipoPost[tmpltName], how='outer', indicator='Where').drop_duplicates()
 		dfInter = dfUnion[dfUnion.Where=='both']
@@ -300,11 +394,13 @@ def searchDiff(datosEquipoPre, datosEquipoPost, dTmplt):
 		dfCompl['Where'] = dfCompl['Where'].str.replace('left_only','Pre')
 		dfCompl['Where'] = dfCompl['Where'].str.replace('right_only','Post')
 
-		countDif[tmpltName] = dfCompl.sort_values(by=dTmplt[tmpltName]['listOfcolumns'])
+		orderedColums = RTR_ID[routerId] + filterCols
+
+		countDif[tmpltName] = dfCompl.sort_values(by = orderedColums)
 
 	return countDif
 
-def findMajor(count_dif, dTmplt):
+def findMajor(count_dif, dTmplt, routerId):
 	#Makes a table from the results of searching for Major errors in the post table define in yml file for specific template, 
 	# or down if is not define the words for the template, which are not in the Pre table
 
@@ -315,7 +411,9 @@ def findMajor(count_dif, dTmplt):
 		df         = pd.DataFrame()
 
 		for majorWord in dTmplt[tmpltName]['majorDown']:
-			
+
+			filterCols = dTmplt[tmpltName]['filterColumns']
+
 			df1 = count_dif[tmpltName][count_dif[tmpltName]['Where']=='Post']
 			
 			if len(df1) > 0:
@@ -324,7 +422,8 @@ def findMajor(count_dif, dTmplt):
 				df1 = pd.DataFrame(columns=count_dif[tmpltName].columns)
 
 			df = pd.concat([df, df1])
-			df = df.sort_values(by=dTmplt[tmpltName]['listOfcolumns'])
+
+			df = df.sort_values(by = RTR_ID[routerId] + filterCols)
 
 		countDown[tmpltName] = df
 
@@ -358,10 +457,10 @@ def constructExcel(df_final, count_dif, searchMajor, folderLog):#Sort the data a
 	writer    = pd.ExcelWriter(fileName, engine='xlsxwriter') #creates instance of an excel workbook
 	workbook  = writer.book
 
-	print('Saving Excel')
-	
 	# Create index tab
 	indexSheet = workbook.add_worksheet('index')
+
+	print('\nSaving Excel')
 
 	for idx,template in enumerate(df_final.keys()):
 
@@ -448,7 +547,7 @@ def constructExcel(df_final, count_dif, searchMajor, folderLog):#Sort the data a
 		if len(dfMajor) > 0:
 			dfMajor.to_excel(writer, sheet_name=sheet_name, startrow=(len(dfData)+(len(dfDiff)))+10, startcol=0)
 		
-		print('#')
+		print('#',idx,template)
 	
 	writer.save() #saves workbook to file in python file directory
 
@@ -462,7 +561,8 @@ def main():
 	parser1.add_argument('-tf', '--templateFolder', type=str, default='Templates/', help='Folder where templates reside. Used both for PRE and POST logs. Default=Templates/')
 	parser1.add_argument('-tf-post', '--templateFolderPost', type=str, default='Templates/', help='If set, use this folder of templates for POST logs. Default=Templates/')
 	parser1.add_argument('-te', '--templateEngine', choices=['ttp','textFSM'], default='textFSM', type=str, help='Engine for parsing. Default=textFSM.')
-	parser1.add_argument('-v'  ,'--version',        help='Version', action='version', version='Saldivar/Aimaretto - (c)2022 - Version: 3.2.3' )
+	parser1.add_argument('-ri', '--routerId', choices=['name','ip','both'], default='name', type=str, help='Router Id to be used within the tables in the Excel report. Default=name.')
+	parser1.add_argument('-v'  ,'--version',        help='Version', action='version', version='Saldivar/Aimaretto - (c)2022 - Version: 3.4.0' )
 
 	args               = parser1.parse_args()
 	preFolder          = args.preFolder
@@ -472,6 +572,7 @@ def main():
 	templateFolder     = args.templateFolder
 	templateEngine     = args.templateEngine
 	templateFolderPost = args.templateFolderPost
+	routerId           = args.routerId
 
 	if _platform == "win64" or _platform == "win32":
 		templateFolder = templateFolder.replace('/', '\\')
@@ -483,7 +584,7 @@ def main():
 		dTmplt = readTemplate(csvTemplate, templateFolder, templateEngine)
 		dLog   = readLog(preFolder, formatJson)
 
-		df_final    = parseResults(dTmplt, dLog, templateFolder, templateEngine)
+		df_final    = parseResults(dTmplt, dLog, templateFolder, templateEngine, routerId)
 		count_dif   = {}
 		searchMajor = {}
 
@@ -523,11 +624,11 @@ def main():
 		dLogPre  = readLog(preFolder, formatJson)
 		dLogPost = readLog(postFolder, formatJson)
 			
-		datosEquipoPre  = parseResults(dTmpltPre,  dLogPre,  templateFolder,     templateEngine)
-		datosEquipoPost = parseResults(dTmpltPost, dLogPost, templateFolderPost, templateEngine)
+		datosEquipoPre  = parseResults(dTmpltPre,  dLogPre,  templateFolder,     templateEngine, routerId)
+		datosEquipoPost = parseResults(dTmpltPost, dLogPost, templateFolderPost, templateEngine, routerId)
 
-		count_dif       = searchDiff(datosEquipoPre, datosEquipoPost, dTmpltPre)
-		searchMajor     = findMajor(count_dif, dTmpltPre)
+		count_dif       = searchDiff(datosEquipoPre, datosEquipoPost, dTmpltPre, routerId)
+		searchMajor     = findMajor(count_dif, dTmpltPre, routerId)
 		df_final        = makeTable(datosEquipoPre, datosEquipoPost)
 
 		constructExcel(df_final, count_dif, searchMajor, postFolder)